requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
attrs==25.4.0
bcrypt==5.0.0
beanie==1.22.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.100.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lazy-model==0.2.0
motor==3.3.2
multidict==6.7.0
passlib==1.7.4
propcache==0.4.1
pyasn1==0.6.1
pycparser==2.23
pydantic==1.10.12
PyJWT==2.7.0
pymongo==4.6.0
python-dotenv==0.21.1
python-jose==3.5.0
python-multipart==0.0.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
toml==0.10.2
typing_extensions==4.15.0
uvicorn==0.23.2
yarl==1.22.0
